Add pileup on specific region and enable bgzipped VCF output
For measures of speed, the pysam pileup function is run on all reads and
the filtering of the sites specified in the newly introduced commandline
option 'region' is performed on the pandas DataFrame of the pileup calls
using its loc function.

Instead of a plain VCF, the biopython sub-module bgzf is used to write
bgzipped VCF output.